requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distro==1.9.0
Django==4.2.13
django-cors-headers==4.3.1
django-countries==7.6.1
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
grpcio==1.65.1
grpcio-health-checking==1.65.1
grpcio-tools==1.65.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
marqo==3.6.0
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
mutagen==1.47.0
networkx==3.2.1
numba==0.60.0
numpy==1.24.4
openai==1.37.0
openai-whisper==20231117
packaging==24.1
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
paypalrestsdk==1.13.3
protobuf==5.27.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.13.1
tbb==2021.13.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.43.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
validators==0.33.0
weaviate-client==4.7.0
websockets==12.0
yt-dlp==2024.7.16
zope.interface==6.4.post2
